Move runSetupRunMounts to run_common.go
